+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 
 ### Enhancements:
 
+- feat(audit-log): add event-mapping command group ([#1875](https://github.com/fastly/cli/pull/1875))
+
 ### Dependencies:
+
 - build(deps): `github.com/rogpeppe/go-internal` from 1.15.0 to 1.16.0 ([#1871](https://github.com/fastly/cli/pull/1871))
 
 ## [v15.5.0](https://github.com/fastly/cli/releases/tag/v15.5.0) (2026-08-04)